chore: refactor to allow unit testing
