+++ pkg/authorizer/suite_test.go
@@ -1,4 +1,4 @@
-package auth
+package authorizer
 
 import (
 	"testing"
