+++ pkg/authorizer/token_retriever.go
@@ -1,0 +1,77 @@
+package authorizer
+
+import (
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"net/http"
+	"net/url"
+)
+
+const (
+	armResource         = "https://management.azure.com/"
+	msiMetadataEndpoint = "http://169.254.169.254/metadata/identity/oauth2/token"
+)
+
+// TokenRetriever is an instance of ManagedIdentityTokenRetriever
+type TokenRetriever struct{}
+
+// AcquireARMToken acquires the managed identity ARM access token
+func (tr *TokenRetriever) AcquireARMToken(clientID string, resourceID string) (AccessToken, error) {
+	msiEndpoint, err := url.Parse(msiMetadataEndpoint)
+	if err != nil {
+		return "", err
+	}
+
+	parameters := url.Values{}
+	if clientID != "" {
+		parameters.Add("client_id", clientID)
+	} else {
+		parameters.Add("mi_res_id", resourceID)
+	}
+
+	parameters.Add("resource", armResource)
+	parameters.Add("api-version", "2018-02-01")
+
+	msiEndpoint.RawQuery = parameters.Encode()
+
+	req, err := http.NewRequest("GET", msiEndpoint.String(), nil)
+	if err != nil {
+		return "", err
+	}
+	req.Header.Add("Metadata", "true")
+
+	client := &http.Client{}
+	var resp *http.Response
+	defer closeResponse(resp)
+
+	resp, err = client.Do(req)
+	if err != nil {
+		return "", fmt.Errorf("failed to send metadata endpoint request: %w", err)
+	}
+
+	if resp.StatusCode != 200 {
+		responseBytes, _ := ioutil.ReadAll(resp.Body)
+		return "", fmt.Errorf("Metadata endpoint returned error status: %d. body: %s", resp.StatusCode, string(responseBytes))
+	}
+
+	responseBytes, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return "", fmt.Errorf("failed to read metadata endpoint response: %w", err)
+	}
+
+	var tokenResp tokenResponse
+	err = json.Unmarshal(responseBytes, &tokenResp)
+	if err != nil {
+		return "", fmt.Errorf("failed to unmarshal metadata endpoint response: %w", err)
+	}
+
+	return AccessToken(tokenResp.AccessToken), nil
+}
+
+func closeResponse(resp *http.Response) {
+	if resp == nil {
+		return
+	}
+	resp.Body.Close()
+}
